* fixing a segfault on higher optimization levels on linux

+++ src/lib/crf/features/feature_word.cc
@@ -34,7 +34,7 @@ namespace NLP {
         WordEntry *next;
 
         static Hash::Hash hash(const char *type, const Word value) {
-          return Hash::Hash((value.index() >> 2)*7 + type);
+          return Hash::Hash((value.index() >> 2)*7 + (uint64_t)type);
         }
 
         static WordEntry *create(Util::Pool *pool, const uint64_t index,
